+++ src/com/skilldistillery/jets/entities/TroopTransport.java
@@ -2,9 +2,11 @@ package com.skilldistillery.jets.entities;
 
 public class TroopTransport extends Jet {
 
-	public TroopTransport(String model, double speed, int range, long price) {
+	public TroopTransport(String model, double speed, double range, double price) {
 		super(model, speed, range, price);
 		
 	}
+
+
 		
 }
